ovpn backport first commit
This commit introduces a tool to generate a backported copy of the ovpn
kernel module for older kernels. Originally, the in-tree ovpn module,
designed to enhance OpenVPN performance, was added to the mainline
repository in version 6.16. To enable this functionality on legacy
systems, we need to compile the module against earlier kernel sources
with some modifications.

Key components:
 - compat-include/: Mimics the kernel’s include directory
   structure and provides header files with the necessary adjustments.
 - linux-compat.h: Includes additional macros required for
   compatibility.
 - ovpn.patch: Contains adjustments to the module source code.
 - backports-ctl.sh: The main script that imports and prepares the
   source files.

NOTE: WIP! Currently compiles up to kernel tag v5.1

Signed-off-by: Ralf Lici <ralf@mandelbit.com>