Fix tstats uninitialized in kernel version < 6.9
On Linux version 6.9.0 and later, the dev->pcpu_stat_type setting
enables automatic initialization of dev->tstats by the netdev core. For
older kernels, however, tstats must be manually allocated and
initialized. This commit addresses the missing initialization on
pre-6.9.0 kernels.

Additionally, this commit includes a backport of the latest changes from
tstats to dstats for kernel versions 6.14 and above.

Reported-by: Qingfang Deng <dqfext@gmail.com>
Closes: https://github.com/OpenVPN/ovpn-backports/issues/1
Signed-off-by: Ralf Lici <ralf@mandelbit.com>