Improve net-next tree handling for `get-ovpn -k` and update repository
When using the `--keep` option with the `get-ovpn` command, the cloned
kernel tree remains after extracting the ovpn source files. However,
running `get-ovpn` again fails do to improper handling of the persistent
net-next directory. This commit resolves the issue by ensuring the
kernel tree is updated as a shallow copy if it already exists.

Additionally, this commit updates the next-next repository source from
`mandelbitdev` to `OpenVPN`.

Signed-off-by: Ralf Lici <ralf@mandelbit.com>