Set specific ovpn-net-next commit pointer and add MODULE_VERSION
We now incorporate a specific ovpn-net-next commit hash into
backports-ctl.sh to ensure that our out-of-tree module is generated
deterministically and reproducibly. This commit pointer will be updated
only after a new ovpn-net-next commit has been tested and confirmed to
work with backports, helping to prevent patch failures as well as
compilation or runtime errors.

Additionally, we append a call to `MODULE_VERSION()` in main.c,
embedding a version string that represents the out-of-tree module. The
version string is composed as follows:
  $tree/$branch-$kernel_version-$backports_commit

Where:
  - $tree: the name of the repository from which the ovpn sources were
    extracted ('ovpn-net-next' for now).
  - $branch: either 'main' or 'development', indicating whether the used
    ovpn-net-next commit originates from the official upstream code or
    from our testing branch. Since the ovpn-net-next repository is
    cloned as a shallow copy, this information is retrieved from
    ovpn-backports, which contains two corresponding branches.
  - $kernel_version: the kernel version of the ovpn-net-next tree used
    to extract the ovpn sources.
  - $backports_commit: the abbreviated commit hash of the current
    ovpn-backports commit that indirectly points to a specific
    ovpn-net-next commit.

For example:
  ovpn-net-next/main-6.15.0-rc4-08b1265

Finally, we remove the get-kernel command, which was primarily used as a
utility during the compilation testing phase. This change simplifies
backports-ctl.sh, ensuring it remains minimal and focused on the
repository's core purpose. However, you can still pass the KERNEL_SRC
variable to `make` to specify a kernel tree different from the one
currently running.

Signed-off-by: Ralf Lici <ralf@mandelbit.com>