Fix version append issue by deferring kernel deletion
When incorporating MODULE_VERSION into the module, the kernel version is
required as part of the version string. However, the kernel tree was
being deleted immediately after extracting the sources, which prevented
the execution of `make kernelversion`. This commit postpones the
deletion of the kernel until the end of the process, ensuring that the
kernel version can be obtained correctly.

Signed-off-by: Ralf Lici <ralf@mandelbit.com>