Align tag format with MODULE_VERSION
We recently introduced a call to MODULE_VERSION() to include backported
module version information. Now we do something similar for the tag to
keep it coherent with the module version. Instead of appending a commit
hash abbreviation (which would be redundant), we append the date to
generate a strictly monotonically increasing tag string. The new tag
format is as follows: $tree/$branch-$kernel_version-$date

Where:
  - $tree: the name of the repository from which the ovpn sources were
    extracted ('ovpn-net-next' for now).
  - $branch: either 'main' or 'development', indicating whether the used
    ovpn-net-next commit originates from the official upstream code or
    from our testing branch. Since the ovpn-net-next repository is
    cloned as a shallow copy, this information is retrieved from
    ovpn-backports, which contains two corresponding branches.
  - $kernel_version: the kernel version of the ovpn-net-next tree used
    to extract the ovpn sources.
  - $date: the current date in the format YYYYMMDD

For example:
  ovpn-net-next/main-6.15.0-rc4-20250514

Signed-off-by: Ralf Lici <ralf@mandelbit.com>