Add ovpn-cli and selftests
Currently only the source files are extracted from the kernel tree. For
testing and debugging purposes we now also extract the ovpn kernel
selftests directory. This not only carries the test scripts but also the
useful "ovpn-cli" tool. We introduce a Makefile target to handle its
compilation and add the selftests files to the list of files to track
when creating or applying the patch.

We also ensure these new files are tracked in the GitHub workflow by
including them in the `git add -f` command.

Signed-off-by: Ralf Lici <ralf@mandelbit.com>