Handle MODULE_VERSION in Makefile
Currently, `MODULE_VERSION()` is dynamically appended to
`drivers/net/ovpn/main.c` using `backports-ctl.sh get-ovpn`. This
results in a hardcoded version string in the source file, which may
inadvertently be included in patches.

To prevent this, we now unconditionally append a call to
`MODULE_VERSION()` in `main.c`, using `OVPN_MODULE_VERSION` as the
argument. During compilation, `OVPN_MODULE_VERSION` is populated with
the correct version information based on the `.version` file, which is
already used to generate the tag.

Signed-off-by: Ralf Lici <ralf@mandelbit.com>