Avoid overwriting .version in source branches
We keep track of version information through the `.version` file, which
is (re)created every time `backports-ctl.sh get-ovpn` is run. Since this
command is also used in the GitHub workflow to generate pre-patched
sources, it ends up overwriting the file with unnecessary data,
complicating versioning.

This commit ensures that we modify the `.version` file only in
non-source branches so that, when building tarballs, the correct version
string is set.

Signed-off-by: Ralf Lici <ralf@mandelbit.com>