Avoid calling kallsyms_lookup_name and reduce unexported symbols usage
On some Intel CPUs, a security feature called Indirect Branch Tracking
(IBT) triggers a kernel BUG when a function is called via a pointer and
lacks an ENDBR instruction at its entry point. This issue was observed
with kallsyms_lookup_name, which we previously accessed via kprobes
during module load. That function pointer was then used to resolve and
cache three unexported kernel symbols related to TCP functionality.

With this change, we adopt a pattern similar to espintcp to eliminate
reliance on two of those symbols: tcpv6_prot and inet6_stream_ops. We
also remove the use of kallsyms_lookup_name entirely. Instead, we now
use kprobes directly to locate the remaining unexported function,
sendmsg_locked, which is typically invoked through a pointer and
therefore assumed to be safe from IBT-related crashes.

Signed-off-by: Ralf Lici <ralf@mandelbit.com>