Limit ovpn_sendmsg_locked definition to v6.5–v6.16 range
The unexported sendmsg_locked function is required only on kernels from
v6.5.0 through v6.16.0. However, our current reimplementation is defined
for all kernels older than v6.16.0, which leads to unused function
warnings on versions prior to v6.5.0.

This commit sets a lower bound of v6.5.0 for the definition, eliminating
the warning and aligning the implementation with actual usage.

Signed-off-by: Ralf Lici <ralf@mandelbit.com>